clean code (to be continued...)

+++ core/src/com/bendk97/postprocessing/effects/LensFlare2.java
@@ -208,7 +208,7 @@ public final class LensFlare2 extends PostProcessorEffect {
 	}
 
 	public int getGhosts () {
-		return (int)lens.getGhosts();
+		return lens.getGhosts();
 	}
 
 	public boolean isBlendingEnabled () {
